refactor: 패키지 구조 변경 (feature 단위)
- config, annotation, model(공통 모델) -> common
- UserRepository, UserService, User -> user
