+++ src/main/java/com/flyby/ramble/common/model/DeviceType.java
@@ -1,10 +1,10 @@
-package com.flyby.ramble.model;
+package com.flyby.ramble.common.model;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 
 @JsonFormat(shape = JsonFormat.Shape.STRING)
-public enum Role {
-    USER,
-    MANAGER,
-    ADMIN
+public enum DeviceType {
+    ANDROID,
+    IOS,
+    WEB
 }
