+++ src/main/java/com/flyby/ramble/user/model/Role.java
@@ -1,10 +1,10 @@
-package com.flyby.ramble.model;
+package com.flyby.ramble.user.model;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 
 @JsonFormat(shape = JsonFormat.Shape.STRING)
-public enum DeviceType {
-    ANDROID,
-    IOS,
-    WEB
+public enum Role {
+    USER,
+    MANAGER,
+    ADMIN
 }
